Version bump to 3.1.3 for Internal Release Squashed commit of the following:
commit 0692fc069dde5cfea1984bbfc8de300404956eba
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Thu Dec 28 08:26:59 2023 +0545

    Tweak - Deprecate ur_has_date_field function

commit 6c15f1de8af964b43cd27cb53b238eb172ee1c16
Author: Milan Chaudhary <109399375+MILAN88888@users.noreply.github.com>
Date:   Wed Dec 27 17:13:51 2023 +0545

    Fix - Select field option not changing instantly (#635)

commit 28ee5200665ae8cba98c84657afc20e4de55b113
Author: Milan Chaudhary <109399375+MILAN88888@users.noreply.github.com>
Date:   Wed Dec 27 17:12:28 2023 +0545

    Fix - Placeholder is not working in the country field (#634)

    * Fix - Placeholder is not working in the country field

    * Add - New option None in the default of country field

    * Fix - PHPcs

commit bc4c82fb0b4e32cc8458cf2fbb33b458d88615b8
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Wed Dec 27 16:50:14 2023 +0545

    UR-1302 Enhance - Smart tags parseable endpoint content (#645)

    * Enhance - Endpoint content made smart tags friendly

    * Tweak - Change in dashboard template

commit 059b23b1371604c047968d24667640d424547e25
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Wed Dec 27 11:56:54 2023 +0545

    Deprecated user_registration_before_edit_profile_form action hook.

    * Fix - Profile completeness hook mismatch issue

    * Deprecated user_registration_before_edit_profile_form action hook

commit bbed6d89c4ad6961d0682a074dd022d9b9b72a18
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Wed Dec 27 11:54:00 2023 +0545

    Tweak - Default form fields re arrangement (#644)

commit 342bf2202d27b5d3e022648d9139d6d490320594
Author: Lekhnath Pandey <lekhnath005@gmail.com>
Date:   Wed Dec 27 10:15:34 2023 +0545

    Trigger hook for redirect after logout

commit faf4381b3a7eaa9c4aedec5690f1104db77a7244
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Tue Dec 26 13:52:28 2023 +0545

    Fix - Manual counting query issue

commit cf57ecf833ded1091435b16e5acecdcde793c261
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Tue Dec 26 13:17:52 2023 +0545

    UR-1304 Fix - Send email change confirmation with templates (#638)

    * Fix - Send email change confirmation with templates

    * Fix - PHPCS

    * Updated

commit 3f945599c7681f0af1466bac8b8ea6b4612e3a0b
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Tue Dec 26 12:44:42 2023 +0545

    Fix - Lost Password Invalid key issue (#642)

commit 8b40f1f32ad67debe9a8b048558fe9a0aa6e74ce
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Tue Dec 26 12:31:52 2023 +0545

    UR-1296 Fix - Support decimal value in number field (#637)

commit eda100357e9669693068f9caf9b1367f9bc1c0b5
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Tue Dec 26 12:13:18 2023 +0545

    Fix - Welcome email sent when profile details updated by admin (#636)

    * Fix - Welcome email sent when profile details updated by admin

    * Fix - PHPCS

commit 8865e8070f94ea2e7e5e71260d534f4db535bfc1
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Tue Dec 26 11:06:16 2023 +0545

    Fix - Pending users meta query optimization (#640)

commit 8ad9260c1ab008499fc38b8db046ace626434496
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Tue Dec 26 10:48:06 2023 +0545

    Fix - Manual user counting replaced with raw sql query for better performance (#641)

commit 3108d02af1c04a2944f213af4ed2f9e95ca1c8c4
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Thu Dec 21 09:20:24 2023 +0545

    Fix - Undefined property

commit 7bf5de8ec2cc3f30c8e1972e685dcf882957c813
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Tue Dec 19 14:38:14 2023 +0545

    UR-1306 Tweak - Parse approval link smart tag properly (#639)

    * Tweak - Parse approval link smart tag when needed

    * Tweak - Escape and made the Approve now text translation ready

commit 319a1ad0c56e3cbb259dd6421885da946bc25610
Author: Lekhnath Pandey <lekhnath005@gmail.com>
Date:   Thu Dec 14 12:37:25 2023 +0545

    Dev - Added filter hook to modify find myaccount shortcode in page

+++ includes/validation/class-ur-form-validation.php
@@ -283,7 +283,7 @@ class UR_Form_Validation extends UR_Validation {
 					$form_data->value = sanitize_textarea_field( $form_data->value );
 					break;
 				case 'number':
-					$form_data->value = intval( $form_data->value );
+					$form_data->value = floatval( $form_data->value );
 					break;
 				case 'nickname':
 				case 'first_name':
